Move files around to be more rational
