--- src/NRules/NRules.Json/Utilities/MethodRecord.cs
@@ -1,16 +1,0 @@
-﻿using System;
-
-namespace NRules.Json.Utilities
-{
-    internal readonly struct MethodRecord
-    {
-        public readonly string Name;
-        public readonly Type DeclaringType;
-
-        public MethodRecord(string name, Type declaringType)
-        {
-            Name = name;
-            DeclaringType = declaringType;
-        }
-    }
-}
